+++ src/FsLexYacc.Build.Tasks/AssemblyInfo.fs
@@ -4,9 +4,9 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FsLexYacc.Build.Tasks")>]
 [<assembly: AssemblyProductAttribute("FsLexYacc.Build.Tasks")>]
 [<assembly: AssemblyDescriptionAttribute("FsLex/FsYacc lexer/parser generation tools")>]
-[<assembly: AssemblyVersionAttribute("6.0.2")>]
-[<assembly: AssemblyFileVersionAttribute("6.0.2")>]
+[<assembly: AssemblyVersionAttribute("6.0.3")>]
+[<assembly: AssemblyFileVersionAttribute("6.0.3")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "6.0.2"
+    let [<Literal>] Version = "6.0.3"