+++ RELEASE_NOTES.md
@@ -4,3 +4,7 @@
 #### 6.0.2 - June 16 2014
 * Logo was added
 * FsLexYacc.Runtime published as a separate NuGet package
+
+#### 6.0.3 - June 18 2014
+* FsLex/FsYacc output redirected to VS Output window
+* FsYacc verbose output added to MSBuild log (and VS Output window)